pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relievo"
version = "0.1.0"
description = "CLI to automate shaded relief map creation in Blender (Huffman method)"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "requests>=2.25",
    "pyproj>=3.0",
    "Pillow>=9.0",
    "tomli>=1.1; python_version < '3.11'",
    # gdal (osgeo) must be installed via conda-forge, NOT pip
]

[project.scripts]
relievo = "relievo.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.4"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
